+++ src/entities/users/index.ts
@@ -1,0 +1,2 @@
+export * from './api'
+export { default as UsersApi } from './api'
