+++ src/router/index.ts
@@ -1,13 +1,13 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import Login from '../views/Login.vue'
-import NovoConvertidosList from '../views/NovoConvertidosList.vue'
-import NovoConvertidoForm from '../views/NovoConvertidoForm.vue'
-import UsuariosList from '../views/UsuariosList.vue'
-import UsuarioForm from '../views/UsuarioForm.vue'
-import UsuarioInactivateConfirm from '../views/UsuarioInactivateConfirm.vue'
-import ContatoNovoConvertidoList from '../views/ContatoNovoConvertidoList.vue'
-import ContatoNovoConvertidoForm from '../views/ContatoNovoConvertidoForm.vue'
-import AcoesContatoForm from '../views/AcoesContatoForm.vue'
+import NovoConvertidosList from '../entities/novos-convertidos/views/NovoConvertidosList.vue'
+import NovoConvertidoForm from '../entities/novos-convertidos/views/NovoConvertidoForm.vue'
+import UsuarioForm from '../entities/users/views/UsuarioForm.vue'
+import UsuarioInactivateConfirm from '../entities/users/views/UsuarioInactivateConfirm.vue'
+import UsuariosList from '../entities/users/views/UsuariosList.vue'
+import ContatoNovoConvertidoList from '../entities/contatos/views/ContatoNovoConvertidoList.vue'
+import ContatoNovoConvertidoForm from '../entities/contatos/views/ContatoNovoConvertidoForm.vue'
+import AcoesContatoForm from '../entities/contatos/views/AcoesContatoForm.vue'
 import Logout from '../views/Logout.vue'
 import { useAuthStore } from '../stores/auth'
